flask_catalog_login/login/web/requirements.txt
flask
pymongo
flask-cors
Flask-And-Redis
